WARNING: use int32 for brn2, hash.c still uses uint32

+++ util.c
@@ -125,7 +125,7 @@ static void __attribute__((format(printf, 1, 2))) error(char *format, ...);
 #endif
 
 #define STRUCT_ARRAY_SIZE(struct_object, ArrayType, array_length) \
-    (uint32)(SIZEOF(*(struct_object)) + (array_length*SIZEOF(ArrayType)))
+    (int64)(SIZEOF(*(struct_object)) + (array_length*SIZEOF(ArrayType)))
 
 #define SWAP(x, y) do { __typeof__(x) SWAP = x; x = y; y = SWAP; } while (0)
 
@@ -476,7 +476,7 @@ qsort64(void *base, int64 n, int64 size,
 }
 
 #if OS_WINDOWS
-static uint32
+static int32
 util_nthreads(void) {
     SYSTEM_INFO sysinfo;
     memset64(&sysinfo, 0, SIZEOF(sysinfo));
@@ -484,9 +484,9 @@ util_nthreads(void) {
     return sysinfo.dwNumberOfProcessors;
 }
 #else
-static uint32
+static int32
 util_nthreads(void) {
-    return (uint32)sysconf(_SC_NPROCESSORS_ONLN);
+    return (int32)sysconf(_SC_NPROCESSORS_ONLN);
 }
 #endif
